use realtime instead of batch
