--- .github/workflows/docs_summarization_check.yml
@@ -1,60 +1,0 @@
-name: Check Docs Summarization
-
-on:
-  push:
-    branches: [release/**]
-
-jobs:
-  check-batch:
-    runs-on: ubuntu-latest
-    permissions:
-      contents: read
-      id-token: write
-      actions: read
-    
-    steps:
-      - uses: actions/checkout@v3
-      
-      - name: Set up Python
-        uses: actions/setup-python@v4
-        with:
-          python-version: '3.11'
-          
-      - name: Install dependencies
-        run: |
-          python -m pip install --upgrade pip
-          pip install openai huggingface_hub
-          
-      - name: List artifacts
-        uses: actions/github-script@v6
-        id: artifacts
-        with:
-          script: |
-            const artifacts = await github.rest.actions.listArtifactsForRepo({
-              owner: context.repo.owner,
-              repo: context.repo.name,
-            });
-            const batchArtifact = artifacts.data.artifacts
-              .find(artifact => artifact.name.startsWith('batch-id-'));
-            if (!batchArtifact) {
-              throw new Error('No batch ID artifact found');
-            }
-            console.log(`Found artifact: ${batchArtifact.name}`);
-            return batchArtifact.name;
-          
-      - name: Download batch ID
-        uses: actions/download-artifact@v3
-        with:
-          name: ${{ steps.artifacts.outputs.result }}
-          
-      - name: Download repomix outputs
-        uses: actions/download-artifact@v3
-        with:
-          name: repomix-outputs
-          path: repomix-outputs
-          
-      - name: Process batch results and upload to HuggingFace
-        env:
-          OPENAI_API_KEY: ${{ secrets.OPENAI_API_KEY }}
-          HF_TOKEN: ${{ secrets.HF_TOKEN }}
-        run: python scripts/check_batch_output.py
